Zaimplementowanie patrolujących wrogów Określenie pozycji i prędkości aktora przez double, przystosowanie movementHandlera do AI, oddzielenie bardziej podstawowych bibliotek do /libs

+++ bomberman/src/actor/actor_essentials.c
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-Actor *createActor( int x, int y, int speed, enum Direction dir, ALLEGRO_BITMAP *bmp )
+Actor *createActor( double x, double y, double speed, enum Direction dir, ALLEGRO_BITMAP *bmp )
 {
     Actor *actor = (Actor *)malloc( sizeof( Actor ) );
